modifica modello

+++ PacMan/src/logicModel/PacMan.java
@@ -14,15 +14,16 @@ import org.newdawn.slick.SlickException;
  */
 public class PacMan extends Giocatore {
     private int vite;
-    private double xPos, yPos;
+    private int xPos, yPos;
     private boolean power;
     
     /*
     TODO: definire le posizioni iniziali di PacMan
     */
-    private final double X_MAIN_POS = 32;
-    private final double Y_MAIN_POS = 32;
+    private final int X_MAIN_POS = 32;
+    private final int Y_MAIN_POS = 32;
     private final int MAIN_VITE = 5;
+    
 
     public PacMan() throws SlickException {
 //        super(new Image[]{new Image("data/pacman0.png"), new Image("data/pacman1.png")});
@@ -33,8 +34,8 @@ public class PacMan extends Giocatore {
 
     
     //getter
-    public double getxPos() {return xPos;}
-    public double getyPos() {return yPos;}
+    public int getxPos() {return xPos;}
+    public int getyPos() {return yPos;}
     public int getVite() {return vite;}
     public boolean isPower() {return power;}
     
@@ -51,8 +52,8 @@ public class PacMan extends Giocatore {
         if(y != 0)
             setYPos(y);
     }
-    private void setXPos(double x){xPos += x;}
-    private void setYPos(double y){yPos += y;}
+    private void setXPos(int x){xPos += x;}
+    private void setYPos(int y){yPos += y;}
